// src/shared/types/cargo.enum.ts
/**
 * Enum para os cargos de usuário no sistema
 * Enum alinhado com as definições do Prisma (sem strings)
 */
export enum CargoUsuario {
  ADMIN = 'ADMIN',
  PROFESSOR = 'PROFESSOR',
  ESPECIALISTA = 'ESPECIALISTA',
  COORDENADOR = 'COORDENADOR',
  DIRETOR = 'DIRETOR',
  ADMINISTRADOR = 'ADMINISTRADOR',
}